input: move function to config
This function is only about rc, so fits better
in config.
