+++ include/config/tablet.h
@@ -22,5 +22,6 @@ enum rotation tablet_parse_rotation(int value);
 uint32_t tablet_button_from_str(const char *button);
 void tablet_button_mapping_add(uint32_t from, uint32_t to);
 void tablet_load_default_button_mappings(void);
+uint32_t tablet_get_mapped_button(uint32_t src_button);
 
 #endif /* LABWC_TABLET_CONFIG_H */
